requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
yarl==1.5.1

pandas~=1.2.4
numpy~=1.20.2